Accept flat networks in the Neutron metadata agent
The Neutron metadata agent implicitly expects nested networks
by requiring that it have either a router ID or a network ID. In
flat networks (like the one set up by Calico), one metadata agent
may serve ports on multiple networks and without OpenStack
routers.

This patch allows the metadata agent (and the namespace proxy) to
function in a network that does not use virtual layer 2 networks,
by relying on the fact that in such an OpenStack deployment IP
addresses will be unique across the data center.